Combine SetupFormSubmitButtons into SetupForm.
