+++ assets/js/modules/tagmanager/components/setup/index.js
@@ -18,5 +18,4 @@
 
 export { default as SetupForm } from './SetupForm';
 export { default as SetupFormInstructions } from './SetupFormInstructions';
-export { default as SetupFormSubmitButtons } from './SetupFormSubmitButtons';
 export { default as SetupMain } from './SetupMain';
